Move error checking code to a sub module
