+++ utils/test_utils/__init__.py
@@ -1,0 +1,4 @@
+from .check_for_errors import check_for_assertion_errors as check_for_assertion_errors
+from .check_for_errors import check_for_errors as check_for_errors
+
+__all__ = ["check_for_assertion_errors", "check_for_errors"]
